Added better support for Safari
Safari doesn't come with Flash by default, so Safari users would
fallback to the flash version and not see the download button at all.

Now, when Safari is detected, it creates a data URI, and makes it the
href of an anchor element set to open in a new tab. Then some
instructions on how to save the URI appear. This provides a flashless
download solution for Safari, whilst working around its behaviour of
opening data URIs directly instead of downloading them.

Updated test page.