Added better callback functions
Replaced the afterInit callback with two new ones - afterFlashInit,
which is called after the Flash component is initialised, and
afterSafarInit, which is called after the Safari only portion is
initialised.

The idea is that you'd use these callbacks to display different help
text, e.g. some text to explain the Safari download, and some to
display for people that don't have Flash Player.